Fix flakiness in BindingApiServiceTest
Test for BindingApiService fails to read the complete response
from the server and doesn't expect a chunked response.

This patch asks the server to close the connection after sending
a response. This makes easier to read the full response without
being flaky and also causes the server to never use chunked
encoding.

Change-Id: Ic9e468cf5e6731b79ba64962dbedb97900890661
Signed-off-by: Adrian Serrano <adrian@midokura.com>