Agent minions: Wait for minions to shutdown before exiting
The Agent companion process where minions run is supposed to
be kept alive and running without exiting (otherwise the main
Agent process will keep restarting it). However, this process
would exit if the started minions doesn't start a non-daemon
thread.

This patch just waits for the minions to be terminated before
exiting the Agent companion process.

Ref: IIOT-556
Change-Id: I1e6136d0fe8acd4b831c409d8454a425d848893d
Signed-off-by: Xavi León <xavi.leon@midokura.com>