Remove PortTranslator.fipArpTableUpdateOps()
This function was intended as a fix for MI-1210, but is incorrect for a couple
of reasons. First, the intent is to handle the update of a router gateway port's
MAC address, which is not allowed by Neutron. Second, this code doesn't actually
run on the update of a router gateway port's MAC address, but on the update of a
VIF port's MAC address. This is allowed, but as the code was designed to be run
on router gateway MAC address update, the behavior is not correct in the case of
a VIF port update, which doesn't actually require any changes to FIP-related ARP
entries.

This patch removes fipArpTableUpdateOps(), a helper function, and a related
test, and reverts changes to neutron_translation.md describing its behavior.

Change-Id: I267cf2eba55c867762410388fe7afe70fb2de3f3
Signed-off-by: Brandon Berg <bberg@midokura.com>
Fixes: MI-1474