Remove ref counting from flow expiration queues
Flow expiration queues used to keep ref counts on the managed flows, as
it was possible to remove a flow early, due to invalidation, and
removing from the expiration queues at that time is expensive.

This patch introduces the concept of a flow id, which is stored in the
expiration queue instead of the reference to the managed flow. The flow
can be looked up using this id. This allows the flow to be removed, and
expiration to occur without the flow being deleted twice.

This change paves the way for removing refcounting on the managed flows
completely, and will allow flow deletion operations to be simplified.

Change-Id: Ie95b77923b9e6b5731563470d6380ac9c7109d36
Ref: MI-2266
Signed-off-by: Ivan Kelly <ivan@midokura.com>