C++ implementation of flow table
Flow table implemented in C++. Still missing many features, like
expiration and invalidation. The flow table is a fixed size. When you
add a flow you get a flow id back. If the table is full you get a null
flow id back. To avoid this, you can evict and old flow before adding if
the table is full.

The flow table is a ring buffer of sorts. Entries are added at a cursor,
which wraps around when it reaches the end of the buffer. It is assumed
that the entry to the left of the cursor is the newest entry in the
table, and the entry to the right is the oldest.

Change-Id: I7ae0d5fb141edfc45f12b24bc052cc41f0dafd40
Ref: MI-2270
Signed-off-by: Ivan Kelly <ivan@midokura.com>