Callback registry for flow removed callbacks
Flow removed callbacks will live for the lifetime of the flow. This is
bad for GC as it means they will likely be tenured and then need to be
collected soon after. To solve problems like this, we are moving the
flow table offheap. However, to move stuff offheap, it needs to be
possible to serialize all data.

This patch introduces a callback registry to heap with serializing
callbacks. Code that adds callbacks will first register a callback once
and get back a callback id. Then to add the callback to the flow, they
add the callback id and a byte array which contains the serialized
arguments to the callback.

When the callbacks are to be called, the id is used to look up the
callback object, and it is called with byte array as an argument. The
callback is itself responsible for deserializing the arguments.

Change-Id: I59d4dcbc3eb83d97b6504a00cdf28592f25cab96
Ref: MI-2267
Signed-off-by: Ivan Kelly <ivan@midokura.com>