Use CallbackRegistry for flow removal callbacks
Use the callback registry for flow removal callbacks. This modifies all
flow removal callbacks to be serializable.

The majority of the patch is changes to pass the callback registry
around.

The interesting changes are in RouterMapper, BridgeMapper and
FlowStateReplicator. Each of these register callbacks and have new code
for serialzation and deserialization of the arguments of those
callbacks.

Change-Id: I4f59578f409489730583521980cc5189ab6fcfc9
Ref: MI-2267
Signed-off-by: Ivan Kelly <ivan@midokura.com>