Only send flow state to the worker that'll use it
The agent generating the flow state knows what the return packet will
look like, so it is also able to generate the connection hash that can
be used to choose which worker thread it is processed on.

This patch generates the connection hash with the flow state and
transmits it to the remote host which then processes it on the correct
worker. The connection hash is encoded in the IP identification +
flags + fragment offset fields, as they are not being used and to
allow backwards compatibility (other fields are checked by previous
midonet versions). Also, if the agent receives legacy flow state,
it should be sent to all workers, as legacy flow state doesn't have
the connection hash encoded in the message.

There are two special cases that needs to be handled:
- when a forward flow generates an error icmp message. The connection
  hash does not correspond to the pre-computed one in the flow state.
  In this case, we need to parse the ICMP data to get the original
  packet header and compute the connection hash.
- when a flow is flow traced. In this case, the flow is send back to
  user space for tracing on an egress host, which means that the
  connection hash is not the inverse, but the forward one. As the
  flow state was sent with the inverse connection hash, we should
  calculate the inverse connection hash for that packet.

Change-Id: I6079f813371f39537b94507a6ec0cc83c55dfba8
Ref: MI-1273
Signed-off-by: Ivan Kelly <ivan@midokura.com>
Signed-off-by: Xavi León <xavi.leon@midokura.com>