AgentReboot: reclaim datapath and ports
Up until now, we were deleting the datapath (and hence all
the ports and flows) upon agent reboot. This patch is a first
step to reuse the existing datapath to minimize the downtime
after a reboot. Specifically, during datapath bootstrap
we check for the existence of the MidoNet datapath instead
of recreating. In the case of ports being created, we check
if they exists and if that's the case we set them to the new
datapath connection, so the netlink upcalls will be processed
by the corresponding handler.

Note: until the out-of-band flow expiration is implemented,
activating this feature implies that existing flows are not
removed.

Change-Id: I57caa447f50fb352bc0564483b111fe9dc8ce621
Ref: MI-2382
Signed-off-by: Xavi León <xavi.leon@midokura.com>