Agent: reduce heap size on templates
According to the latest benchmarks, we can reduce significantly
the size of the JVM heap size. The reference ratio now is:

Heap Size (GB) = flow rate (kfps) / 2.5

This size is computed so the number of garbage collections
are kept to a minimum without reducing it too much as to force
tenuring of short lived objects.

Change-Id: I6438289ca76858d944119b4c6502320bb2d3f7ee
Ref: MI-2338
Signed-off-by: Xavi León <xavi.leon@midokura.com>