NSDB: import object and state cache
This patch open sources the object and state cache from
Insights so it can be reused by MidoNet. Some data structures
have been adapted to accomodate the specifics of the
MidoNet project. In summary, this implements a cache
for NSDB objects (topology and state). Upon start,
an exposed observable starts emitting notifications with
the current NSDB state. This will be later used to create
a snapshot upon request from a Midonet Agent.

NOTE: this set of classes might be subject to change to adapt
to the specific scenario of speeding up the agent bootstrap.
However, it should be left untouched as much as possible to
not prevent its reuse by Insights if needed.

Change-Id: Ie6b12d9bbd07a76ec610b8403e04da6d42d53a17
Ref: MI-2551
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
Signed-off-by: Alexandre Fonseca <alex.fonseca@midokura.com>
Signed-off-by: Adrian Serrano <adrian@midokura.com>
Signed-off-by: Mario Macías <mario@midokura.com>
Signed-off-by: Xavi León <xavi.leon@midokura.com>